Update multiple tests to use the new ofsted record

+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Services/AcademyServiceTests.cs
@@ -43,14 +43,14 @@ public class AcademyServiceTests
         var academies = new[]
         {
             new AcademyOfsted("1", "Academy 1", new DateTime(2022, 12, 1),
-                new OfstedRating(OfstedRatingScore.Good, new DateTime(2023, 1, 1)),
-                new OfstedRating(OfstedRatingScore.RequiresImprovement, new DateTime(2023, 2, 1))),
+                new OfstedRating((int)OfstedRatingScore.Good, new DateTime(2023, 1, 1)),
+                new OfstedRating((int)OfstedRatingScore.RequiresImprovement, new DateTime(2023, 2, 1))),
             new AcademyOfsted("2", "Academy 2", new DateTime(2022, 11, 2),
-                new OfstedRating(OfstedRatingScore.Good, new DateTime(2023, 1, 2)),
-                new OfstedRating(OfstedRatingScore.RequiresImprovement, new DateTime(2023, 3, 1))),
+                new OfstedRating((int)OfstedRatingScore.Good, new DateTime(2023, 1, 2)),
+                new OfstedRating((int)OfstedRatingScore.RequiresImprovement, new DateTime(2023, 3, 1))),
             new AcademyOfsted("3", "Academy 3", new DateTime(2022, 10, 3),
-                new OfstedRating(OfstedRatingScore.Good, new DateTime(2023, 1, 3)),
-                new OfstedRating(OfstedRatingScore.RequiresImprovement, new DateTime(2023, 4, 1)))
+                new OfstedRating((int)OfstedRatingScore.Good, new DateTime(2023, 1, 3)),
+                new OfstedRating((int)OfstedRatingScore.RequiresImprovement, new DateTime(2023, 4, 1)))
         };
 
         _mockAcademyRepository.Setup(a => a.GetAcademiesInTrustOfstedAsync(uid))
